requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
apify-client>=1.6.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
webdriver-manager>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

